DC-5139: Maintenance and upgrades

+++ app/uk/gov/hmrc/ups/controllers/UpdatedOn.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 HM Revenue & Customs
+ * Copyright 2023 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,7 +28,7 @@ import scala.concurrent.{ ExecutionContext, Future }
 import scala.math.BigDecimal.RoundingMode
 import play.api.mvc.Results._
 import uk.gov.hmrc.mongo.MongoComponent
-import uk.gov.hmrc.time.DateTimeUtils
+import uk.gov.hmrc.ups.utils.DateTimeUtils
 
 import javax.inject.Inject
 
@@ -49,9 +49,9 @@ class UpdatedOn @Inject()(mongoComponent: MongoComponent, counterRepository: Mon
         val offset = optOffset.getOrElse(1)
         for {
           count   <- upsRepository.count()
-          updates <- upsRepository.find(offset, limit.value)
+          updates <- upsRepository.find(offset.toLong, limit.value)
         } yield {
-          val pages: Int = (BigDecimal(count) / BigDecimal(limit.value)).setScale(0, RoundingMode.UP).intValue()
+          val pages: Long = (BigDecimal(count) / BigDecimal(limit.value)).setScale(0, RoundingMode.UP).longValue
           Ok(
             Json.toJson(
               UpdatedPrintPreferences(
